pyproject.toml
[project]
name = "agents-team"
version = "0.1.0"
description = "Install one canonical team of AI agents into Codex, Claude Code, and OpenCode."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0.2",
    "typer>=0.12.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
]

[project.scripts]
agents-team = "agents_team.cli:app"

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["agents_team*"]
